% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_assignDS.R
\name{adonis_assignDS}
\alias{adonis_assignDS}
\title{Generates the results of adonis function and store it on the server}
\usage{
adonis_assignDS(
  formula,
  data.name.transmit,
  permutations = 999,
  method = "bray",
  strata = NULL,
  contr.unordered = "contr.sum",
  contr.ordered = "contr.poly"
)
}
\arguments{
\item{formula}{an parameter of unspecified purpose}

\item{data.name.transmit}{a string character, the name of a dataframe}

\item{permutations}{an parameter of unspecified purpose}

\item{method}{an parameter of unspecified purpose}

\item{strata}{an parameter of unspecified purpose}

\item{contr.unordered}{an parameter of unspecified purpose}

\item{contr.ordered}{an parameter of unspecified purpose}
}
\value{
the results of \code{adonisDS} function
}
\description{
This function is implemented by wapper around the \code{vegan} function \code{adonisDS}.
}
\details{
The function returns ...
}
\author{
Stuart Wheater for DataSHIELD Development Team
}
