% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegdist_assignDS.R
\name{vegdist_assignDS}
\alias{vegdist_assignDS}
\title{Generates the results of vegdist function and store it on the server}
\usage{
vegdist_assignDS(
  matrix.name.transmit,
  method = "bray",
  binary = FALSE,
  diag = FALSE,
  upper = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{matrix.name.transmit}{a string character, the name of a dataframe}

\item{method}{an parameter of unspecified purpose}

\item{binary}{an parameter of unspecified purpose}

\item{diag}{an parameter of unspecified purpose}

\item{upper}{an parameter of unspecified purpose}

\item{na.rm}{an parameter of unspecified purpose}
}
\value{
the results of \code{vegdistDS} function
}
\description{
This function is implemented by wapper around the \code{vegan} function \code{vegdistDS}.
}
\details{
The function returns ...
}
\author{
Stuart Wheater for DataSHIELD Development Team
}
